+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Use --net=host mode for Docker in tests. Extend restart tests.
 - Fix termination issue in Process tests
 - Bump Go version (1.22.8) and dependencies for CVE fixes
+- Allow to sync setup.json in default mode (no host and port specified)
 
 ## [v0.18.10](https://github.com/arangodb-helper/arangodb/tree/0.18.10) (2024-09-26)
 - Fix for member discovery during leader election